Document diagnostic mechanisms and mission profiles
